Rename demo pkg to just demo-pkg because it's not just for core anymore
