+++ packages/demo-pkg/package.json
