+++ packages/demo-pkg/srcts/index.ts
@@ -1,0 +1,8 @@
+// Simple bindings without web components or react
+import "./plain-bindings";
+
+// React bindings
+import "./react-components";
+
+// Web component bindings (using lit)
+import "./webcomponents";
