+++ packages/demo-pkg/tsconfig.json
